pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[project]
name = "nmdc-ai-eval"
description = "Framework and data for performing evaluations for AI-powered NMDC tools."
authors = [
  {name = "Harry Caufield", email = "jhc@lbl.gov"},
  {name = "Mark Andrew Miller", email = "MAM@lbl.gov"},
]
license = "BSD-3-Clause"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dynamic = ["version"]

dependencies = [
  "click<8.3",
  "ipython>=9.10.0",
  "llm-claude-3>=0.11",
  "llm-gemini>=0.29",
  "llm-matrix>=0.1.3",
  "oaklib>=0.6.0",
  "pandas>=2.0",
  "pyyaml>=6.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "ruff>=0.11",
  "pre-commit>=3.3",
  "mypy>=1.19.1",
  "deptry>=0.24.0",
  "pip-audit>=2.10.0",
  "pytest-cov>=7.0.0",
  "pandas-stubs>=3.0.0.260204",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nmdc_ai_eval"]

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
fallback-version = "0.0.0"

[tool.uv]
constraint-dependencies = [
  "anthropic>=0.84.0",    # llm-claude-3 runtime dep
  "linkml>=1.10.0",       # linkml <1.10 references Format.JSON removed in linkml-runtime 1.10
  "setuptools>=70.0,<81", # oaklib→eutils uses pkg_resources, removed in setuptools 82
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short --durations=5"
filterwarnings = [
  "ignore::DeprecationWarning:sssom_schema",
  "ignore::pandas.errors.Pandas4Warning:sssom",
]
markers = ["api: tests requiring LLM API calls (excluded from CI)"]

[tool.coverage.run]
omit = ["src/nmdc_ai_eval/run_suite.py"]  # API-only; untestable without live LLM calls

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # isort
  "W",   # pycodestyle warnings
  "B",   # bugbear
  "S",   # bandit (security)
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S603", "S607"]  # assert, subprocess in tests
"datasets/**" = ["S311"]  # pseudo-random OK for sampling eval data

[tool.codespell]
skip = [
  "LICENSE",
  "pyproject.toml",
  "uv.lock",
  "datasets/*",
]
quiet-level = 3

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.deptry]
per_rule_ignores = {DEP002 = ["ipython", "llm-claude-3", "llm-gemini", "pyyaml"], DEP003 = ["llm"]}

[tool.typos.files]
extend-exclude = [
  "LICENSE",
  "uv.lock",
  "pyproject.toml",
  "datasets/*",
]
